+++ src/world/DynamicObject.cpp
@@ -145,7 +145,7 @@ void DynamicObject::UpdateTargets()
             if (GetDistanceSq(target) <= radius)
             {
                 pAura = sSpellFactoryMgr.NewAura(m_spellProto, m_aliveDuration, u_caster, target, true);
-                for (uint32 i = 0; i < 3; ++i)
+                for (uint8 i = 0; i < 3; ++i)
                 {
                     if (m_spellProto->Effect[i] == SPELL_EFFECT_PERSISTENT_AREA_AURA)
                     {
