+++ src/world/QuestMgr.h
@@ -156,7 +156,7 @@ class SERVER_DECL QuestMgr : public Singleton <QuestMgr>
 
         inline int32 QuestHasMob(Quest* qst, uint32 mob)
         {
-            for (uint32 i = 0; i < 4; ++i)
+            for (uint8 i = 0; i < 4; ++i)
                 if (qst->required_mob[i] == (int32)mob)
                     return qst->required_mobcount[i];
             return -1;
@@ -164,7 +164,7 @@ class SERVER_DECL QuestMgr : public Singleton <QuestMgr>
 
         inline int32 GetOffsetForMob(Quest* qst, uint32 mob)
         {
-            for (uint32 i = 0; i < 4; ++i)
+            for (uint8 i = 0; i < 4; ++i)
                 if (qst->required_mob[i] == (int32)mob)
                     return i;
 
@@ -173,7 +173,7 @@ class SERVER_DECL QuestMgr : public Singleton <QuestMgr>
 
         inline int32 GetOffsetForItem(Quest* qst, uint32 itm)
         {
-            for (uint32 i = 0; i < MAX_REQUIRED_QUEST_ITEM; ++i)
+            for (uint8 i = 0; i < MAX_REQUIRED_QUEST_ITEM; ++i)
                 if (qst->required_item[i] == itm)
                     return i;
